Fixes bug in LOD which results in no frustum cull for geometry within an LOD
This change fixes a bug in LOD which results in geometry within an LOD never being culled. You can see this in the scene stats diagram because the DRW and TRI values do not reduce to zero when camera is facing away from all geometry when LOD nodes are present in the scene. This fix requires the changes in the previous pull request for Object3D which allows the reuse of the calculateChildVisibility and collectChildDraws functions by the LOD class and thus allow the equivalent code (or rather, the nearly equivalent code) to be removed from LOD class. 

This change significantly improved frame rates for my scenes which make heavy use of LOD.